+++ summoner-cli/test/Test/Golden.hs
@@ -91,6 +91,9 @@ fullProject = Settings
     , settingsStylish        = Just "This is stylish-haskell.yaml\n"
     , settingsContributing   = Just "This is contributing guide\n"
     , settingsNoUpload       = True
+    , settingsFiles          =
+        [ File "extra.txt" [text|See full content of the file [here](@github)|]
+        ]
     }
   where
     mitLicense :: License
@@ -148,6 +151,7 @@ smallProject = Settings
     , settingsStylish        = Nothing
     , settingsContributing   = Nothing
     , settingsNoUpload       = True
+    , settingsFiles          = mempty
     }
 
 -- Orphan instances