+++ summoner-tui/src/Summoner/Tui.hs
@@ -29,12 +29,13 @@ import System.Directory (doesDirectoryExist, doesFileExist, getCurrentDirectory,
 
 import Summoner.Ansi (errorMessage, infoMessage)
 import Summoner.CLI (Command (..), NewOpts (..), ShowOpts (..), getFinalConfig, runScript, summon)
+import Summoner.Config (ConfigP (cFiles))
 import Summoner.Decision (Decision (..))
 import Summoner.Default (defaultConfigFile)
 import Summoner.GhcVer (showGhcMeta)
 import Summoner.License (License (..), LicenseName, fetchLicense, parseLicenseName,
                          showLicenseWithDesc)
-import Summoner.Project (initializeProject)
+import Summoner.Project (fetchSources, initializeProject)
 import Summoner.Text (alignTable)
 import Summoner.Tui.Field (disabledAttr)
 import Summoner.Tui.Form (KitForm, SummonForm (..), getCurrentFocus, isActive, mkForm, recreateForm)
@@ -70,11 +71,13 @@ summonTuiNew :: NewOpts -> IO ()
 summonTuiNew newOpts@NewOpts{..} = do
     -- configure initial state for TUI application
     finalConfig <- getFinalConfig newOpts
+    files <- fetchSources newOptsOffline (cFiles finalConfig)
     configFilePath <- findConfigFile
     let initialKit = configToSummonKit
             newOptsProjectName
             newOptsOffline
             configFilePath
+            files
             finalConfig
 
     -- run TUI app
@@ -227,6 +230,7 @@ drawNew dirs kitForm = case kit ^. shouldSummon of
             [ informationBlock
             , validationBlock
             , configBlock
+            , deprecationBlock
             , fill ' '
             ]
       where
@@ -240,6 +244,9 @@ drawNew dirs kitForm = case kit ^. shouldSummon of
         infoTxt :: Text -> Widget SummonForm
         infoTxt = withAttr "blue-fg" . txtWrap . (<>) " ⓘ  "
 
+        deprecationTxt :: Text -> Widget SummonForm
+        deprecationTxt = withAttr "yellow-fg" . txtWrap . (<>) " ⚠  "
+
         validationBlock :: Widget SummonForm
         validationBlock = vBox $ case formErrorMessages dirs kitForm of
             []     -> [withAttr "green-fg" $ str " ✔  Project configuration is valid"]
@@ -249,6 +256,13 @@ drawNew dirs kitForm = case kit ^. shouldSummon of
         configBlock = case kit ^. configFile of
             Nothing   -> emptyWidget
             Just file -> infoTxt $ toText file <> " file is used"
+
+        deprecationBlock :: Widget SummonForm
+        deprecationBlock = case (kit ^. stylish, kit ^. contributing) of
+            (Nothing, Nothing) -> emptyWidget
+            (s, c) -> vBox $
+                [ deprecationTxt "Option 'stylish' is deprecated, use 'files'" | isJust s ]
+             ++ [ deprecationTxt "Option 'contributing' is deprecated, use 'files'" | isJust c]
 
     help, helpBody :: Widget SummonForm
     help     = borderLabel "Help" (helpBody <+> fill ' ')
@@ -353,11 +367,12 @@ theMap = attrMap V.defAttr
     , (editFocusedAttr,         V.black `Brick.on` V.white)
     , (invalidFormInputAttr,    V.white `Brick.on` V.red)
     , (focusedFormInputAttr,    V.black `Brick.on` V.yellow)
-    , (listSelectedAttr,        V.black  `Brick.on` V.cyan)
+    , (listSelectedAttr,        V.black `Brick.on` V.cyan)
     , (listSelectedFocusedAttr, V.black `Brick.on` V.white)
     , (disabledAttr,            fg V.brightBlack)
     , ("blue-fg",               fg V.blue)
     , ("green-fg",              fg V.green)
+    , ("yellow-fg",             fg V.yellow)
     , ("red-fg",                fg V.brightRed)
     , (borderAttr,              fg V.cyan)
     , ("tree",                  fg V.cyan)